+++ src/rules/index.js
@@ -2,6 +2,8 @@ import atExtendNoMissingPlaceholder from "./at-extend-no-missing-placeholder"
 import atFunctionPattern from "./at-function-pattern"
 import atImportNoPartialExtension from "./at-import-no-partial-extension"
 import atImportNoPartialLeadingUnderscore from "./at-import-no-partial-leading-underscore"
+import atImportPartialExtensionBlacklist from "./at-import-partial-extension-blacklist"
+import atImportPartialExtensionWhitelist from "./at-import-partial-extension-whitelist"
 import atMixinNoArgumentlessCallParentheses from "./at-mixin-no-argumentless-call-parentheses"
 import atMixinPattern from "./at-mixin-pattern"
 import dollarVariableNoMissingInterpolation from "./dollar-variable-no-missing-interpolation"
@@ -16,6 +18,8 @@ export default {
   "at-function-pattern": atFunctionPattern,
   "at-import-no-partial-extension": atImportNoPartialExtension,
   "at-import-no-partial-leading-underscore": atImportNoPartialLeadingUnderscore,
+  "at-import-partial-extension-blacklist": atImportPartialExtensionBlacklist,
+  "at-import-partial-extension-whitelist": atImportPartialExtensionWhitelist,
   "at-mixin-no-argumentless-call-parentheses": atMixinNoArgumentlessCallParentheses,
   "at-mixin-pattern": atMixinPattern,
   "dollar-variable-no-missing-interpolation": dollarVariableNoMissingInterpolation,