+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Added: `partial-no-import` rule.
 - Added: `media-feature-value-dollar-variable` rule.
+- Added: `at-import-partial-extension-blacklist` rule.
+- Added: `at-import-partial-extension-whitelist` rule.
+- Deprecated: `at-import-no-partial-extension` rule.
 
 # 1.1.1